+++ inst/app_template/R/example-module.R
@@ -1,5 +1,5 @@
-mymoduleUI <- function(id, label = "Counter") {
-  # Al uses of Shiny input/output IDs in the UI must be namespaced,
+exampleModuleUI <- function(id, label = "Counter") {
+  # All uses of Shiny input/output IDs in the UI must be namespaced,
   # as in ns("x").
   ns <- NS(id)
   tagList(
@@ -8,7 +8,7 @@ mymoduleUI <- function(id, label = "Counter") {
   )
 }
 
-mymoduleServer <- function(id) {
+exampleModuleServer <- function(id) {
   # moduleServer() wraps a function to create the server component of a
   # module.
   moduleServer(